+++ frontend/src/components/cards/MarkdownCard.tsx
@@ -4,8 +4,10 @@ import remarkGfm from "remark-gfm";
 import rehypeRaw from "rehype-raw";
 import rehypeHighlight from "rehype-highlight";
 import CopyButton from "../buttons/CopyButton";
+import { ImagePreviewModal } from "../inputs/ImagePreviewModal";
+import useImageHook from "@/hooks/useImageHook";
 
-const MarkdownCard = ({ content }: { content: string }) => {
+const BaseCard = ({ content }: { content: string }) => {
     return (
         <ReactMarkdown
             components={{
@@ -101,6 +103,39 @@ const MarkdownCard = ({ content }: { content: string }) => {
             {content}
         </ReactMarkdown>
     );
+}
+
+const MarkdownCard = ({ content }: { content: string | any[] }) => {
+    const { handleImageClick, handleImageClear, previewImage, previewImageIndex } = useImageHook();
+
+    // If string, return a single card
+    if (typeof content === "string") {
+        return <BaseCard content={content} />;
+    }
+    // If array, return a list of cards
+    if (Array.isArray(content)) {
+        return content.map((item, index) => {
+            if (item.type === "text") {
+                return <BaseCard content={item.text} key={index} />;
+            }
+            if (item.type === "image_url") {
+                return (
+                    <div className="flex justify-center items-center py-1" key={index}>
+                        <img src={item.image_url.url} alt="image" className="cursor-pointer" onClick={() => handleImageClick(item.image_url.url, index)} />
+                        {previewImage && (
+                            <ImagePreviewModal
+                                image={previewImage}
+                                index={previewImageIndex}
+                                onClose={() => handleImageClear()}
+                            />
+                        )}
+                    </div>
+                );
+            }
+        });
+    }
+    console.error("Invalid content");
+    return null;
 };
 
 export default MarkdownCard;
