+++ frontend/src/components/drawers/ThreadHistoryDrawer.tsx
@@ -80,7 +80,7 @@ export function ThreadHistoryDrawer({ isOpen, onClose }: ThreadHistoryDrawerProp
                           >
                               <div className="w-full pr-8">
                                   <p className="text-sm font-medium line-clamp-2 max-w-60">
-                                      {thread.messages[1]?.content ? truncateFrom(thread.messages[1].content, 'end', "...", 70) : "New Chat"}
+                                      {thread.messages[1]?.content && typeof(thread.messages[1]?.content) === "string" ? truncateFrom(thread.messages[1].content, 'end', "...", 70) : "New Chat"}
                                   </p>
                                   <p className="text-xs text-muted-foreground mt-1 truncate">
                                       {formatDistanceToNow(new Date(thread.ts), {
